+++ CHANGELOG.md
@@ -15,3 +15,5 @@ The log below details the changes during development of this version:
 * 2025-04-23: Fix in JSON-schemas: Changed `main` property in Graphics manifest to be **mandatory**.
   (Before, it was defined as mandatory in the specification document, but not in the JSON-schemas.)
 * 2025-05-16: Add `renderRequirements` property to Graphics manifest
+* 2025-05-16: Add `data` argument to the `load` method.
+  Before, the `data`-payload was only sent using the `updateData()` method. Now it must be sent on `load()` as well.